infradash/requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
boto3>=1.34.0
google-auth>=2.28.0
google-cloud-bigquery>=3.17.0
google-cloud-billing>=1.13.0
oci>=2.120.0
requests>=2.31.0
python-dotenv>=1.0.0
